api/requirements.txt
requests
beautifulsoup4
python-telegram-bot==13.7
flask
urllib3